Create events ext, prepare jams cog for file split
